app/requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.1
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.1
Django==3.0.1
django-filter==2.2.0
djangorestframework==3.11.0
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
mccabe==0.6.1
more-itertools==8.0.2
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
